[binance] rename tests class names
